GetSystem improved to avoid AV detection
Since named pipes impersonation is one of the Meterpreter's default getsystem methods, a lot of AV systems are aware of this technique and they tag it as malicious as soon as you try to use it. In order to avoid AV detection, the GetSystem functionality will try to impersonate a System process's access token and only if that fails it will use the old named pipes impersonation method.